Update issue 1577
Adding a 2.4.5 updates text file.  I based this off of all the issues I
could find that had been completed since the last release, but please
feel free to add or alter things in this list before the final release.

I know that there have been some changes to InVEST-2, for example, that
are not yet in this release notes.